fix: switch get_state return value to uint8

+++ smart_contracts/common/abi_types.py
@@ -24,7 +24,7 @@ class ProposalTypedGlobalState(arc4.Struct):
     vote_open_ts: arc4.UInt64
     status: arc4.UInt64
     funding_category: arc4.UInt64
-    focus: arc4.UInt64
+    focus: arc4.UInt8
     funding_type: arc4.UInt64
     requested_amount: arc4.UInt64
     locked_amount: arc4.UInt64
