+++ smart_contracts/proposal/contract.py
@@ -907,7 +907,7 @@ class Proposal(
             vote_open_ts=arc4.UInt64(self.vote_open_ts.value),
             status=arc4.UInt64(self.status.value),
             funding_category=arc4.UInt64(self.funding_category.value),
-            focus=arc4.UInt64(self.focus.value),
+            focus=arc4.UInt8(self.focus.value),
             funding_type=arc4.UInt64(self.funding_type.value),
             requested_amount=arc4.UInt64(self.requested_amount.value),
             locked_amount=arc4.UInt64(self.locked_amount.value),
